+++ RELEASE_NOTES.md
@@ -1,2 +1,7 @@
-#### 0.0.1.2-alpha - August 12 2015
+#### 0.0.1.5-alpha - September 21 2015
+* new binary serializer
+* new shared memory stream
+* Include deployment scripts for windows domain clusters
+
+#### 0.0.1.3-alpha - August 12 2015
 * Initial Setup at GitHub